requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
langchain==0.3.20
langchain-community==0.3.18
langchain-google-genai==2.1.3
faiss-cpu==1.9.0.post1
pypdf==5.4.0
pydantic==2.10.6
python-dotenv==1.0.1
tiktoken==0.8.0
google-generativeai==0.8.3
numpy<2
